requirements.txt
streamlit
python-dotenv
qdrant-client
requests
pandas
numpy
sentence-transformers
groq
